raspi4: Remove raspi-firmware package
This package is only meant for non-UEFI booting Raspberry Pi boards as
it expects that you're using the non-UEFI boot mode and disk layout
which the original Raspberry Pis required (ie: /boot/firmware is its own
partition on your disk and houses the kernel and a pile of firmware
which comes from this package).

Since we're not using that boot mode (UEFI is superior) we do not need
this package.  Additionally, having this package causes issues because
we do not want a /boot/firmware partition, so it actively works against
us.

Fixes: https://github.com/bradfa/debos-configs/issues/1